rename: registerSchema > Schema 폴더명 변경
