+++ src/components/pages/Signup.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 import { SignForm } from '../molecules';
-import { SignupSchemaType, signupSchema } from '../../registerSchema';
+import { SignupSchemaType, signupSchema } from '../../schema';
 import { registerUser } from '../../services/auth';
 import { FirebaseError } from 'firebase/app';
